Reverts caching of Tag.namespace
This is a revert of a2c7d55de7db849b22b23aa69fd9fd5cf02e3edd that would
lead to crashes when documents have a funky mixture of several default
namespaces.
Due to that removal the hashing of Namespace objects isn't needed
anymore and get removed as well.
Where the funk is, the element wrapper cache comes around to bang.
Bless the funk when the effing cache is gone.