integrations-tests: Adds a test that verifies TagNode.location_path
With each run a randomly selected sample of data from the corpora is
tested.
The inclusion as integration test is motivated by the fact that a
representative coverage in the unit tests would take too much time.
At the same time, there are no relevant criteria to curate a subset
of documents for testing and such curation would have to be repeated
for future extensions of the corpora.

A future interface for running these tests could include
configuration options to set the sample sizes.