Removes deprecated, namespace related features and behaviour
Generally the internal representation of empty/null prefixes and
namespaces in declarations is unambiguously an empty string. While
users can alternatively pass in `None` to declare a default
namespace.

Methods that consider namespace declarations default to use the node
instance's namespace as default.
Also common declarations of namespaces that are defined in W3C
recommendations are used as fallback.

The opportunity to express a delb specific evaluation behaviour in a
more elaborate way was seized. One can still miss the subtleties.
test_document::test_css_select was discovered to be wrong in that
regard.