Fixes _reduce_whitespace_between_siblings
where also the item selection based on equality for removal from a
sequence was wrong.
As it surfaced by manifesting different results of whitespace reduction
depending on the container type, a test that shall ensure that each
usage yields the same was added beside a sample document from the
integration tests suite.